+++ src/TTTR.cpp
@@ -536,7 +536,7 @@ size_t TTTR::get_number_of_records_by_file_size(
 
 void ranges_by_time_window(
         int **output, int *n_output,
-        uint64_t *input, uint32_t n_input,
+        unsigned long long *input, int n_input,
         double minimum_window_length,
         double maximum_window_length,
         int minimum_number_of_photons_in_time_window,
